+++ CopilotKit/examples/next-pages-router/CHANGELOG.md
@@ -1,19 +1,124 @@
 # next-pages-router
 
-## 1.4.0-pre-1-4-0.0
-
-### Minor Changes
-
-- Add LangGraph Cloud support
+## 1.3.16-mme-lgc-langgraph-package.9
 
 ### Patch Changes
 
+- update entry
 - Updated dependencies
-  - @copilotkit/runtime@1.4.0-pre-1-4-0.0
-  - @copilotkit/react-core@1.4.0-pre-1-4-0.0
-  - @copilotkit/react-textarea@1.4.0-pre-1-4-0.0
-  - @copilotkit/react-ui@1.4.0-pre-1-4-0.0
-  - @copilotkit/shared@1.4.0-pre-1-4-0.0
+  - @copilotkit/react-textarea@1.3.16-mme-lgc-langgraph-package.9
+  - @copilotkit/react-core@1.3.16-mme-lgc-langgraph-package.9
+  - @copilotkit/react-ui@1.3.16-mme-lgc-langgraph-package.9
+  - @copilotkit/runtime@1.3.16-mme-lgc-langgraph-package.9
+  - @copilotkit/shared@1.3.16-mme-lgc-langgraph-package.9
+
+## 1.3.16-mme-lgc-langgraph-package.8
+
+### Patch Changes
+
+- update entry
+- Updated dependencies
+  - @copilotkit/react-textarea@1.3.16-mme-lgc-langgraph-package.8
+  - @copilotkit/react-core@1.3.16-mme-lgc-langgraph-package.8
+  - @copilotkit/react-ui@1.3.16-mme-lgc-langgraph-package.8
+  - @copilotkit/runtime@1.3.16-mme-lgc-langgraph-package.8
+  - @copilotkit/shared@1.3.16-mme-lgc-langgraph-package.8
+
+## 1.3.16-mme-lgc-langgraph-package.7
+
+### Patch Changes
+
+- update entry in tsup config
+- Updated dependencies
+  - @copilotkit/react-textarea@1.3.16-mme-lgc-langgraph-package.7
+  - @copilotkit/react-core@1.3.16-mme-lgc-langgraph-package.7
+  - @copilotkit/react-ui@1.3.16-mme-lgc-langgraph-package.7
+  - @copilotkit/runtime@1.3.16-mme-lgc-langgraph-package.7
+  - @copilotkit/shared@1.3.16-mme-lgc-langgraph-package.7
+
+## 1.3.16-mme-lgc-langgraph-package.6
+
+### Patch Changes
+
+- Update exports
+- Updated dependencies
+  - @copilotkit/react-textarea@1.3.16-mme-lgc-langgraph-package.6
+  - @copilotkit/react-core@1.3.16-mme-lgc-langgraph-package.6
+  - @copilotkit/react-ui@1.3.16-mme-lgc-langgraph-package.6
+  - @copilotkit/runtime@1.3.16-mme-lgc-langgraph-package.6
+  - @copilotkit/shared@1.3.16-mme-lgc-langgraph-package.6
+
+## 1.3.16-mme-lgc-langgraph-package.5
+
+### Patch Changes
+
+- update tsup config
+- Updated dependencies
+  - @copilotkit/react-textarea@1.3.16-mme-lgc-langgraph-package.5
+  - @copilotkit/react-core@1.3.16-mme-lgc-langgraph-package.5
+  - @copilotkit/react-ui@1.3.16-mme-lgc-langgraph-package.5
+  - @copilotkit/runtime@1.3.16-mme-lgc-langgraph-package.5
+  - @copilotkit/shared@1.3.16-mme-lgc-langgraph-package.5
+
+## 1.3.16-mme-lgc-langgraph-package.4
+
+### Patch Changes
+
+- Update exports
+- Updated dependencies
+  - @copilotkit/react-textarea@1.3.16-mme-lgc-langgraph-package.4
+  - @copilotkit/react-core@1.3.16-mme-lgc-langgraph-package.4
+  - @copilotkit/react-ui@1.3.16-mme-lgc-langgraph-package.4
+  - @copilotkit/runtime@1.3.16-mme-lgc-langgraph-package.4
+  - @copilotkit/shared@1.3.16-mme-lgc-langgraph-package.4
+
+## 1.3.16-mme-lgc-langgraph-package.3
+
+### Patch Changes
+
+- export langchain module
+- Updated dependencies
+  - @copilotkit/react-textarea@1.3.16-mme-lgc-langgraph-package.3
+  - @copilotkit/react-core@1.3.16-mme-lgc-langgraph-package.3
+  - @copilotkit/react-ui@1.3.16-mme-lgc-langgraph-package.3
+  - @copilotkit/runtime@1.3.16-mme-lgc-langgraph-package.3
+  - @copilotkit/shared@1.3.16-mme-lgc-langgraph-package.3
+
+## 1.3.16-mme-sdk-js.2
+
+### Patch Changes
+
+- Ensure intermediate state config is sent as snake case
+- Updated dependencies
+  - @copilotkit/react-textarea@1.3.16-mme-sdk-js.2
+  - @copilotkit/react-core@1.3.16-mme-sdk-js.2
+  - @copilotkit/react-ui@1.3.16-mme-sdk-js.2
+  - @copilotkit/runtime@1.3.16-mme-sdk-js.2
+  - @copilotkit/shared@1.3.16-mme-sdk-js.2
+
+## 1.3.16-mme-sdk-js.1
+
+### Patch Changes
+
+- Update lockfile
+- Updated dependencies
+  - @copilotkit/react-textarea@1.3.16-mme-sdk-js.1
+  - @copilotkit/react-core@1.3.16-mme-sdk-js.1
+  - @copilotkit/react-ui@1.3.16-mme-sdk-js.1
+  - @copilotkit/runtime@1.3.16-mme-sdk-js.1
+  - @copilotkit/shared@1.3.16-mme-sdk-js.1
+
+## 1.3.16-mme-sdk-js.0
+
+### Patch Changes
+
+- Export LangGraph functions
+- Updated dependencies
+  - @copilotkit/runtime@1.3.16-mme-sdk-js.0
+  - @copilotkit/react-core@1.3.16-mme-sdk-js.0
+  - @copilotkit/react-textarea@1.3.16-mme-sdk-js.0
+  - @copilotkit/react-ui@1.3.16-mme-sdk-js.0
+  - @copilotkit/shared@1.3.16-mme-sdk-js.0
 
 ## 1.3.15