+++ CopilotKit/packages/shared/CHANGELOG.md
@@ -1,10 +1,64 @@
 # @copilotkit/shared
 
-## 1.4.0-pre-1-4-0.0
+## 1.3.16-mme-lgc-langgraph-package.9
 
-### Minor Changes
+### Patch Changes
 
-- Add LangGraph Cloud support
+- update entry
+
+## 1.3.16-mme-lgc-langgraph-package.8
+
+### Patch Changes
+
+- update entry
+
+## 1.3.16-mme-lgc-langgraph-package.7
+
+### Patch Changes
+
+- update entry in tsup config
+
+## 1.3.16-mme-lgc-langgraph-package.6
+
+### Patch Changes
+
+- Update exports
+
+## 1.3.16-mme-lgc-langgraph-package.5
+
+### Patch Changes
+
+- update tsup config
+
+## 1.3.16-mme-lgc-langgraph-package.4
+
+### Patch Changes
+
+- Update exports
+
+## 1.3.16-mme-lgc-langgraph-package.3
+
+### Patch Changes
+
+- export langchain module
+
+## 1.3.16-mme-sdk-js.2
+
+### Patch Changes
+
+- Ensure intermediate state config is sent as snake case
+
+## 1.3.16-mme-sdk-js.1
+
+### Patch Changes
+
+- Update lockfile
+
+## 1.3.16-mme-sdk-js.0
+
+### Patch Changes
+
+- Export LangGraph functions
 
 ## 1.3.15